Reorganize comm structure
