+++ casper/src/main/scala/coop/rchain/casper/Casper.scala
@@ -13,6 +13,8 @@ import coop.rchain.crypto.codec.Base16
 import coop.rchain.p2p.Network.ErrorHandler
 import coop.rchain.p2p.effects._
 import coop.rchain.rholang.interpreter.Runtime
+import coop.rchain.comm.transport._
+import coop.rchain.comm.discovery._
 import coop.rchain.shared.AtomicSyncVar
 
 import scala.annotation.tailrec
