+++ casper/src/test/scala/coop/rchain/casper/HashSetCasperTestNode.scala
@@ -9,9 +9,10 @@ import coop.rchain.casper.util.comm.TransportLayerTestImpl
 import coop.rchain.casper.util.comm.CommUtil.casperPacketHandler
 import coop.rchain.comm._
 import coop.rchain.metrics.Metrics
-import coop.rchain.p2p.effects.{PacketHandler, TransportLayer}
+import coop.rchain.p2p.effects.PacketHandler
 import coop.rchain.p2p.EffectsTestInstances._
 import coop.rchain.p2p.Network.dispatch
+import coop.rchain.comm.transport._
 
 import java.nio.file.Files
