+++ node/src/main/scala/coop/rchain/node/grpc.scala
@@ -25,6 +25,8 @@ import java.io.{Reader, StringReader}
 import coop.rchain.casper.api.BlockAPI
 import coop.rchain.node.diagnostics.{JvmMetrics, NodeMetrics, StoreMetrics}
 import coop.rchain.rholang.interpreter.errors.InterpreterError
+import coop.rchain.comm.transport._
+import coop.rchain.comm.discovery._
 
 object GrpcServer {
